Initial sanitized copy-paste
need to add comments later
also need to find out a better solution to create a timeout to the lookups.  socket.settimeout(1) throws errors.  For now have the "cached" entry there to help get passed multiple lookups on the previous lookup.  Maybe put this into some object to search to avoid future lookups on known addresses that do not resolve.